+++ routes/v2/guardians/guardians.js
@@ -106,7 +106,7 @@ router.route('/register')
       if (arbimonService.isEnabled) {
         const idToken = req.headers.authorization
         const arbimonSite = await arbimonService.createSite({
-          ...dbStream.toJSON(),
+          ...dbStream,
           latitude: 0,
           longitude: 0,
           altitude: 0